Add files via upload
This is a bullshit fantasy setting I'm trying to write for shits and giggles. If you've somehow found this and I haven't linked it to you, what the fuck are you doing here? Creeping? Please don't, its not very well written and I'm using GPT4 to bounce ideas around as it grows, so this is a place to upload the document to allow it to actually read and keep up.

So yeah, that.